Docs config update to fix latex output
The Immaterial theme overrides the default Sphinx admonitions to provide extended functionality by default. This overriding breaks the styling in the latex builder, so it must be disabled.
A side effect to this is, that most of the admonition styles break in the html, but this can be fixed with explicit `:class:` markup.
See: https://github.com/jbms/sphinx-immaterial/issues/443